+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Z80:
 * New macro Z80::MathInt::Macros#mul8_signed.
 * New macro Z80::MathInt::Macros#mul_signed9.
 * New macro Z80::MathInt::Macros#mul_signed9_24.
+* New macro Z80::MathInt::Macros#sll8_16.
+* New macro Z80::MathInt::Macros#divmod16_8.
 * Saved 1 cycle on macro Z80::MathInt::Macros#mul.
 * New macro Z80::MathInt::Macros#sign_extend.
 * New options :k_int24 and :optimize for Z80::MathInt::Macros#mul8_24, saved 2 cycles.